Add static value for args in cli tests
