+++ starport/templates/typed/stargate/messages/x/{{moduleName}}/client/cli/tx{{TypeName}}_test.go.plush
@@ -19,6 +19,7 @@ func TestCreate<%= title(TypeName) %>(t *testing.T) {
 	val := net.Validators[0]
 	ctx := val.ClientCtx
 
+    fields := []string{<%= for (field) in Fields { %> "<%= genValidArg(field.DatatypeName) %>", <% } %>}
 	for _, tc := range []struct {
 		desc string
 		args []string
@@ -37,7 +38,10 @@ func TestCreate<%= title(TypeName) %>(t *testing.T) {
 	} {
 		tc := tc
 		t.Run(tc.desc, func(t *testing.T) {
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdCreate<%= title(TypeName) %>(), tc.args)
+            args := []string{}
+            args = append(args, fields...)
+            args = append(args, tc.args...)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdCreate<%= title(TypeName) %>(), args)
 			if tc.err != nil {
 				require.ErrorIs(t, err, tc.err)
 			} else {
@@ -56,12 +60,16 @@ func TestUpdate<%= title(TypeName) %>(t *testing.T) {
 	val := net.Validators[0]
 	ctx := val.ClientCtx
 
-	args := []string{
+    fields := []string{<%= for (field) in Fields { %> "<%= genValidArg(field.DatatypeName) %>", <% } %>}
+	common := []string{
 		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
 		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
 		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
 		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(net.Config.BondDenom, sdk.NewInt(10))).String()),
 	}
+    args := []string{}
+	args = append(args, fields...)
+	args = append(args, common...)
 	_, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdCreate<%= title(TypeName) %>(), args)
 	require.NoError(t, err)
 
@@ -75,18 +83,21 @@ func TestUpdate<%= title(TypeName) %>(t *testing.T) {
 		{
 			desc: "valid",
 			id:   "0",
-			args: args,
+			args: common,
 		},
 		{
 			desc: "key not found",
 			id:   "1",
-			args: args,
+			args: common,
 			code: sdkerrors.ErrKeyNotFound.ABCICode(),
 		},
 	} {
 		tc := tc
 		t.Run(tc.desc, func(t *testing.T) {
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdUpdate<%= title(TypeName) %>(), append([]string{tc.id}, tc.args...))
+            args := []string{tc.id}
+            args = append(args, fields...)
+            args = append(args, tc.args...)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdUpdate<%= title(TypeName) %>(), args)
 			if tc.err != nil {
 				require.ErrorIs(t, err, tc.err)
 			} else {
